Integrated artifact upload to the deploy flow

+++ deploy/team-space-deploy-aws/src/main/java/org/teamspace/deploy/service/impl/DeployServiceImpl.java
@@ -19,7 +19,9 @@ import org.teamspace.network.service.NetworkManager;
 
 import java.io.File;
 
+import static org.teamspace.commons.constants.DeploymentConstants.ARTIFACT_EXTENSION;
 import static org.teamspace.commons.constants.DeploymentConstants.BUCKET_ENTITY_TYPE;
+import static org.teamspace.commons.constants.DeploymentConstants.OVERRIDE_EXISTING_ARTIFACT;
 
 /**
  * Created by shpilb on 06/05/2017.
@@ -49,6 +51,7 @@ public class DeployServiceImpl implements DeployService{
         log.info("Started deploy to region: " + deployRequest
                 .getRegion() + " , env tag: " + deployRequest.getEnvTag());
         initAwsContext(deployRequest.getRegion());
+        uploadArtifact(deployRequest.getArtifactName(), deployRequest.getEnvTag());
         CreateNetworkRequest createNetworkRequest = new CreateNetworkRequest(deployRequest.getEnvTag());
         CreateNetworkResponse createNetworkResponse = networkManager.createNetwork(createNetworkRequest);
         CreateInstanceRequest createInstanceRequest =
@@ -79,16 +82,35 @@ public class DeployServiceImpl implements DeployService{
                 .getRegion() + " , env tag: " + undeployRequest.getEnvTag());
     }
 
-    public void uploadArtifact(String artifactName, Regions region, String envTag){
-        String bucketName = AwsEntitiesHelperUtil.getEntityName(envTag, BUCKET_ENTITY_TYPE).toLowerCase();
-        AmazonS3 s3Client = AwsContext.getS3Client();
-        if(!s3Client.doesBucketExist(bucketName)) {
-            CreateBucketRequest createBucketRequest = new CreateBucketRequest(bucketName, region.getName());
-            s3Client.createBucket(createBucketRequest);
+    public void uploadArtifact(String artifactName, String envTag){
+        log.info("Started artifact upload for artifact: " + artifactName);
+        log.debug("Override flag is: " + OVERRIDE_EXISTING_ARTIFACT);
+        if(OVERRIDE_EXISTING_ARTIFACT) {
+            String fullArtifactName = artifactName + ARTIFACT_EXTENSION;
+            String bucketName = AwsEntitiesHelperUtil.getEntityName(envTag, BUCKET_ENTITY_TYPE).toLowerCase();
+            AmazonS3 s3Client = AwsContext.getS3Client();
+            if (!s3Client.doesBucketExist(bucketName)) {
+                log.debug("Creating bucket: " + bucketName);
+                CreateBucketRequest createBucketRequest = new CreateBucketRequest(bucketName, AwsContext.getRegion().getName());
+                s3Client.createBucket(createBucketRequest);
+            } else {
+                ListObjectsRequest listObjectsRequest = new ListObjectsRequest();
+                listObjectsRequest.withBucketName(bucketName);
+                ObjectListing objectListing = s3Client.listObjects(listObjectsRequest);
+                boolean isArtifactExists = objectListing.getObjectSummaries()
+                        .stream().filter(o -> o.getKey().equals(fullArtifactName)).findFirst().isPresent();
+                if (isArtifactExists) {
+                    log.debug("Deleting artifact: " + fullArtifactName + "from bucket: " + bucketName);
+                    DeleteObjectRequest deleteObjectRequest = new DeleteObjectRequest(bucketName, fullArtifactName);
+                    s3Client.deleteObject(deleteObjectRequest);
+                }
+            }
+            File file = new File(artifactsDir + "/" + fullArtifactName);
+            s3Client.putObject(new PutObjectRequest(
+                    bucketName, fullArtifactName, file));
+            log.debug("Uploaded artifact: " + fullArtifactName + "to bucket: " + bucketName);
         }
-        File file = new File(artifactsDir + "/" + artifactName);
-        s3Client.putObject(new PutObjectRequest(
-                bucketName, artifactName, file));
+        log.info("Finished artifact upload for artifact: " + artifactName);
     }
 
     private void initAwsContext(String region){
